Resolve Issue 47 (#48)
* Dealing with issue 47

Made changes in register so you can add a username. You log in now with a genertic string aka 'username' doesn't check if it's an e-mail anymore

* Clean up so I can pull request

* Login is back to only accepting email

* Dealing with issue 47

Made changes in register so you can add a username. You log in now with a genertic string aka 'username' doesn't check if it's an e-mail anymore

* Clean up so I can pull request

* Login is back to only accepting email

* Api changes

Changed username to email

* Add movement creation error alert.

* Requested Changes

* Remove changes to package-lock.json

Co-authored-by: Robin A. Dorstijn <robin@gridt.org>