Update CI.yml
We would like to make absolutely sure that master is testing positive and pushes are rare, so this is not costly. Pull requests to development are what is most important, but when they succeed that will be the new state of the development branch, so we won't need to test that as well.